+++ backend/api/urls.py
@@ -21,11 +21,6 @@ router.register(r"data_manager", views.DataManagerPermissionViewSet, basename="d
 router.register(r"algo", views.AlgoViewSet, basename="algo")
 router.register(r"algo", views.AlgoPermissionViewSet, basename="algo")
 router.register(r"task", views.ComputeTaskViewSet, basename="task")
-router.register(r"traintuple", views.ComputeTaskViewSet, basename="traintuple")
-router.register(r"predicttuple", views.ComputeTaskViewSet, basename="predicttuple")
-router.register(r"testtuple", views.ComputeTaskViewSet, basename="testtuple")
-router.register(r"aggregatetuple", views.ComputeTaskViewSet, basename="aggregatetuple")
-router.register(r"composite_traintuple", views.ComputeTaskViewSet, basename="composite_traintuple")
 router.register(r"compute_plan", views.ComputePlanViewSet, basename="compute_plan")
 router.register(r"compute_plan_metadata", views.ComputePlanMetadataViewSet, basename="compute_plan_metadata")
 router.register(r"news_feed", views.NewsFeedViewSet, basename="news_feed")
@@ -38,13 +33,6 @@ task_profiling_router.register(r"step", views.TaskProfilingStepViewSet, basename
 
 compute_plan_router = routers.NestedDefaultRouter(router, r"compute_plan", lookup="compute_plan")
 compute_plan_router.register(r"task", views.CPTaskViewSet, basename=f"{CP_BASENAME_PREFIX}task")
-compute_plan_router.register(r"traintuple", views.CPTaskViewSet, basename=f"{CP_BASENAME_PREFIX}traintuple")
-compute_plan_router.register(r"aggregatetuple", views.CPTaskViewSet, basename=f"{CP_BASENAME_PREFIX}aggregatetuple")
-compute_plan_router.register(
-    r"composite_traintuple", views.CPTaskViewSet, basename=f"{CP_BASENAME_PREFIX}composite_traintuple"
-)
-compute_plan_router.register(r"predicttuple", views.CPTaskViewSet, basename=f"{CP_BASENAME_PREFIX}predicttuple")
-compute_plan_router.register(r"testtuple", views.CPTaskViewSet, basename=f"{CP_BASENAME_PREFIX}testtuple")
 compute_plan_router.register(r"algos", views.CPAlgoViewSet, basename=f"{CP_BASENAME_PREFIX}algo")
 compute_plan_router.register(r"perf", views.CPPerformanceViewSet, basename=f"{CP_BASENAME_PREFIX}perf")